remove FileLock trait, merge into FileExt
move code from "guard" module to root
